Trying to break known CPU patterns
In systems with a well known CPU usage patterns (like for example systems with lots of cron jobs running each 10 seconds) there is the risk that the check is perfectly in sync with the CPU peaks and the metric results to be strongly biased towards the CPU usage pattern.
With some randomness we hope to break the pattern.